[level 3] Title: 입국심사, Time: 107.58 ms, Memory: 86.7 MB -BaekjoonHub

+++ 프로그래머스/3/43238. 입국심사/입국심사.java
@@ -7,7 +7,8 @@ class Solution {
         
         // 이분탐색 -> 시간을 축으로
         long left = 0;
-        long right = times[times.length-1] * (long)n; // 최대 시간: 제일 오래걸리는 작업 시간 * 사람 수
+        // 최대 시간: 제일 오래걸리는 작업 시간 * 사람 수 
+        long right = (long)times[times.length-1] * (long)n; 
         long mid=0L, minTime=0L;
         
         while(left<=right){
